changed model of db and ts and started QRModule
