+++ package.json
@@ -36,6 +36,7 @@
     "rxjs-compat": "^6.6.7",
     "sweetalert": "^2.1.2",
     "tslib": "^2.3.0",
+    "uuid": "^9.0.0",
     "zone.js": "~0.12.0"
   },
   "devDependencies": {
