+++ src/app/shared/models/course.model.ts
@@ -1,78 +1,40 @@
-// export interface CourseDto {
-//     code: string;
-//     name: string;
-//     tutor: string;
-//     room:number;
-//     day:string;
-//     documentation: string;
-
-// }
-
-type CloseState = 'No Finalizado' | '[0-15]' | '[15-30]' | 'Más de 30';
-type Confirmation = 'SI' | 'NO';
 
 export interface CourseDto {
-    code: String | undefined;
-    name: String | undefined;
-    start: string | number | null | undefined;
-    end: string | number | null | undefined;
-    preStart: string | number | null | undefined;
-    preEnd: string | number | null | undefined;
-    endDate: string | number | null | undefined;
-    place: string | null | undefined;
-    province: string | null | undefined;
-    solicitudes: number | string | undefined;
-    enrollments: number | string | undefined;
-    realized: number | null | string | undefined;
-    passed: number | null | string | undefined;
-    acreditation: Confirmation | string | undefined;
-    expedientNum: String | null | undefined;
-    creditNum: number | null | undefined;
-    daysToClose: number | null | string | undefined;
-    closeState: CloseState | string | undefined;
-  };
+    code: string;
+    name: string;
+    tutor: string;
+    room:number;
+    day:string;
+    documentation: string;
+}
   
 export class Course {
+    private id: string;
     private code: String | undefined;
     private name: String | undefined;
-    private start: string | number | null | undefined;
-    private end: string | number | null | undefined;
-    private preStart: string | number | null | undefined;
-    private preEnd: string | number | null | undefined;
-    private endDate: string | number | null | undefined;
-    private place: string | null | undefined;
-    private province: string | null | undefined;
-    private solicitudes: number | string | undefined;
-    private enrollments: number | string | undefined;
-    private realized: number | null | string | undefined;
-    private passed: number | null | string | undefined;
-    private acreditation: Confirmation | string | undefined;
-    private expedientNum: String | null | undefined;
-    private creditNum: number | null | undefined;
-    private daysToClose: number | null | string | undefined;
-    private closeState: CloseState | string | undefined;
+    private tutor: string | number | null | undefined;
+    private room: string | number | null | undefined;
+    private day: string | number | null | undefined;
+    private documentation: string | number | null | undefined;
 
-
-    constructor({ code,name,start,end,preStart,preEnd,endDate,place,province,solicitudes,enrollments,realized,passed,acreditation,expedientNum,creditNum,daysToClose,closeState }: CourseDto
+    constructor({ code,name,tutor,room,day,documentation }: CourseDto
     ) {
+       this.id = this.uuidv4(4);
        this.code = code;
        this.name = name;
-       this.start = start;
-       this.end = end;
-       this.preStart = preStart;
-       this.preEnd = preEnd;
-       this.endDate = endDate;
-       this.place = place;
-       this.province = province;
-       this.solicitudes = solicitudes;
-       this.enrollments = enrollments;
-       this.realized = realized;
-       this.passed = passed;
-       this.acreditation = acreditation;
-       this.expedientNum = expedientNum;
-       this.creditNum = creditNum;
-       this.daysToClose = daysToClose;
-       this.closeState = closeState;
+       this.tutor = tutor;
+       this.room = room;
+       this.day = day;
+       this.documentation = documentation;
     }
+    uuidv4 (parts: number): string {
+        const stringArr = [];
+        for(let i = 0; i< parts; i++){
+
+            const S4 = (((1 + Math.random()) * 0x10000) | 0).toString(16).substring(1);
+          stringArr.push(S4);
+        }
+        return stringArr.join('-');
+      }
 }
