+++ src/app/views/excel/excel.component.html
@@ -45,27 +45,19 @@
                 <table id="dtHorizontal" cellspacing="0" class="table table-striped table-bordered table-sm" *ngIf="tableCourse">
                   <thead>
                     <tr scope="row">
+                      <td scope="col">Id</td>
                       <td scope="col">Código</td>
                       <td scope="col">Nombre</td>
-                      <td scope="col">Inicio</td>
-                      <td scope="col">Final</td>
-                      <td scope="col">preInicio</td>
-                      <td scope="col">preFinal</td>
-                      <td scope="col">Fecha de finalización</td>
-                      <td scope="col">Lugar</td>
-                      <td scope="col">Provincia</td>
-                      <td scope="col">Solicitudes</td>
-                      <td scope="col">Matriculados</td>
-                      <td scope="col">Realizados</td>
-                      <td scope="col">Aprobados</td>
-                      <td scope="col">Acreditados</td>
-                      <td scope="col">Nº Expediente</td>
-                      <td scope="col">Nº Credito</td>
-                      <td scope="col">Dias para cierre</td>
-                      <td scope="col">Estado de cierre</td>
+                      <td scope="col">Tutor</td>
+                      <td scope="col">Aula</td>
+                      <td scope="col">Dia de inicio</td>
+                      <td scope="col">Documentacion URL</td>
                     </tr>
                   </thead>
                       <tr  *ngFor="let item of data" scope="row">
+                        <td scope="col" class="item-col">
+                          {{item.id}}
+                        </td>
                         <td scope="col" class="item-col">
                           {{item.code}}
                         </td>
@@ -73,52 +65,16 @@
                           {{item.name}}
                         </td>
                         <td scope="col" class="item-col">
-                          {{item.start}}
+                          {{item.tutor}}
                         </td>
                         <td scope="col" class="item-col">
-                          {{item.end}}
+                          {{item.room}}
                         </td>
                         <td scope="col" class="item-col">
-                          {{item.preStart}}
+                          {{item.day}}
                         </td>
                         <td scope="col" class="item-col">
-                          {{item.preEnd}}
-                        </td>
-                        <td scope="col" class="item-col">
-                          {{item.endDate}}
-                        </td>
-                        <td scope="col" class="item-col">
-                          {{item.place}}
-                        </td>
-                        <td scope="col" class="item-col">
-                          {{item.province}}
-                        </td>
-                        <td scope="col" class="item-col">
-                          {{item.solicitudes}}
-                        </td>
-                        <td scope="col" class="item-col">
-                          {{item.enrollments}}
-                        </td>
-                        <td scope="col" class="item-col">
-                          {{item.realized}}
-                        </td>
-                        <td scope="col" class="item-col">
-                          {{item.passed}}
-                        </td>
-                        <td scope="col" class="item-col">
-                          {{item.acreditation}}
-                        </td>
-                        <td scope="col" class="item-col">
-                          {{item.expedientNum}}
-                        </td>
-                        <td scope="col" class="item-col">
-                          {{item.creditNum}}
-                        </td>
-                        <td scope="col" class="item-col">
-                          {{item.daysToClose}}
-                        </td>
-                        <td scope="col" class="item-col">
-                          {{item.closeState}}
+                          {{item.documentation}}
                         </td>
                       </tr>
              </table>
